kedro-exercises/spaceflight/src/requirements.txt
black==v19.10b0
flake8>=3.7.9, <4.0
ipython~=7.0
isort>=4.3.21, <5.0
jupyter~=1.0
jupyter_client>=5.1,<7.0
jupyterlab==0.31.1
kedro[pandas.CSVDataSet,pandas.ExcelDataSet]==0.16.6
kedro-viz~=3.1
nbstripout==0.3.3
pytest-cov~=2.5
pytest-mock>=1.7.1, <2.0
pytest~=5.0
python-pptx==0.6.18
wheel==0.32.2
scikit-learn==0.23.2
